requirements.txt
#### Richarduino Python Dependencies ####
altgraph==0.17.2
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
future==0.18.2
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pefile==2021.9.3
ping3==3.0.2
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.0
pyserial==3.5
pywin32-ctypes==0.2.0
requests==2.25.1
soupsieve==2.1
urllib3==1.26.2
Werkzeug==2.0.2
windows-curses==2.3.0